- Fixed (I'm pretty sure, at least) Issue where certain circumstances could cause UI Elements to be the wrong size for the specified zoom level

+++ src/main/java/Settings/UserSettings.java
@@ -20,7 +20,7 @@ public class UserSettings {
         Preferences prefs = Preferences.userNodeForPackage(Main.class);
 
         String userTheme = prefs.get("userTheme", "Light");
-        int userZoom = Integer.parseInt(prefs.get("userZoom", "100"));
+        float userZoom = Float.parseFloat(prefs.get("userZoom", "100"));
 
         try {
             UIManager.setLookAndFeel(Themes.getTheme(userTheme));
